requirements.txt
pandas==2.2.3
fastapi==0.115.2
python-multipart==0.0.12
pymysql==1.1.1
vyper-config==1.0.0
